+++ RedPandaIDE/settingsdialog/compilersetoptionwidget.cpp
@@ -88,6 +88,8 @@ static void loadCompilerSetSettings(Settings::PCompilerSet pSet, Ui::CompilerSet
     ui->txtCustomLinkParams->setEnabled(pSet->useCustomLinkParams());
     ui->chkAutoAddCharset->setChecked(pSet->autoAddCharsetParams());
     ui->chkStaticLink->setChecked(pSet->staticLink());
+    ui->spinMaxObjectSize->setValue(pSet->maxObjectSize());
+    ui->chkWarnLargeObject->setChecked(pSet->warnLargeObject());
     //rest tabs in the options widget
 
     ui->optionTabs->resetUI(pSet,pSet->compileOptions());
@@ -120,7 +122,7 @@ static void loadCompilerSetSettings(Settings::PCompilerSet pSet, Ui::CompilerSet
     }
 
     ui->txtPreprocessingSuffix->setText(pSet->preprocessingSuffix());
-    ui->txtCompilationSuffix->setText(pSet->compilationProperSuffix());
+    ui->txtCompilationSuffix->setText(pSet->warnBigObject());
     ui->txtExecutableSuffix->setText(pSet->executableSuffix());
     switch(pSet->compilationStage()) {
     case Settings::CompilerSet::CompilationStage::PreprocessingOnly:
@@ -199,6 +201,8 @@ void CompilerSetOptionWidget::saveCurrentCompilerSet()
     pSet->setCustomLinkParams(ui->txtCustomLinkParams->toPlainText().trimmed());
     pSet->setAutoAddCharsetParams(ui->chkAutoAddCharset->isChecked());
     pSet->setStaticLink(ui->chkStaticLink->isChecked());
+    pSet->setMaxObjectSize(ui->spinMaxObjectSize->value());
+    pSet->setWarnLargeObject(ui->chkWarnLargeObject->isChecked());
 
     pSet->setCCompiler(ui->txtCCompiler->text().trimmed());
     pSet->setCppCompiler(ui->txtCppCompiler->text().trimmed());